refactor(accelerator): extract server::prove core from server.rs (Q2 5/n)
Move the /prove handler + resolve_version + compute_threads + StatusGuard +
MAX_BODY_SIZE into server/prove.rs (pub(crate) where the router/tests reach
them). server.rs keeps the shared ProveError/ProveErrorBody/json_error and is
now thin router+start+health. Router wires post(prove::prove); the 6 helper
unit tests import via super::prove. Behaviour-preserving: 127 lib tests green,
including the prove_* router net and resolve_version/compute_threads units.
